requirements.txt
flask
flask_socketio
requests
pillow
qrcode
